+++ src/Globals/Variables.php
@@ -2,7 +2,6 @@
 
 namespace Statamic\Globals;
 
-use BadMethodCallException;
 use Statamic\Contracts\Data\Augmentable;
 use Statamic\Contracts\Data\Augmented;
 use Statamic\Contracts\Data\Localization;
@@ -14,10 +13,8 @@ use Statamic\Data\HasAugmentedInstance;
 use Statamic\Data\HasOrigin;
 use Statamic\Events\GlobalVariablesBlueprintFound;
 use Statamic\Facades;
-use Statamic\Facades\Compare;
 use Statamic\Facades\Site;
 use Statamic\Facades\Stache;
-use Statamic\Fields\Value;
 use Statamic\GraphQL\ResolvesValues;
 use Statamic\Support\Traits\FluentlyGetsAndSets;
 
@@ -178,18 +175,5 @@ class Variables implements Contract, Localization, Augmentable, ResolvesValuesCo
     public function fresh()
     {
         return Facades\GlobalSet::find($this->id())->in($this->locale);
-    }
-
-    public function __call($method, $args)
-    {
-        $value = $this->augmentedValue($method);
-
-        $value = $value instanceof Value ? $value->value() : $value;
-
-        if (Compare::isQueryBuilder($value)) {
-            return $value;
-        }
-
-        throw new BadMethodCallException(sprintf('Call to undefined method %s::%s()', static::class, $method));
     }
 }
