+++ src/Taxonomies/LocalizedTerm.php
@@ -2,7 +2,6 @@
 
 namespace Statamic\Taxonomies;
 
-use BadMethodCallException;
 use Facades\Statamic\View\Cascade;
 use Illuminate\Contracts\Support\Responsable;
 use Illuminate\Support\Carbon;
@@ -19,7 +18,6 @@ use Statamic\Data\TracksQueriedColumns;
 use Statamic\Exceptions\NotFoundHttpException;
 use Statamic\Facades;
 use Statamic\Facades\Blink;
-use Statamic\Facades\Compare;
 use Statamic\Facades\Site;
 use Statamic\Fields\Value;
 use Statamic\GraphQL\ResolvesValues;
@@ -466,18 +464,5 @@ class LocalizedTerm implements Term, Responsable, Augmentable, Protectable, Reso
     public function fresh()
     {
         return Facades\Term::find($this->id())->in($this->locale);
-    }
-
-    public function __call($method, $args)
-    {
-        $value = $this->augmentedValue($method);
-
-        $value = $value instanceof Value ? $value->value() : $value;
-
-        if (Compare::isQueryBuilder($value)) {
-            return $value;
-        }
-
-        throw new BadMethodCallException(sprintf('Call to undefined method %s::%s()', static::class, $method));
     }
 }
